Crate iceberg_cli

Introduction

Iceberg CLI (iceberg-cli) is a small command line utility that runs SQL queries against tables, which is backed by the DataFusion engine.

Structs

IcebergCatalogList

Constants

ICEBERG_CLI_VERSION